Moving processMostWatched to utils
